+++ bellows/src/loader/file/filesystem.rs
@@ -1,0 +1,155 @@
+use core::ffi::c_void;
+use core::ptr;
+use petroleum::common::{BellowsError, EFI_FILE_INFO_GUID, EfiBootServices, EfiFile, EfiFileInfo, EfiMemoryType, EfiStatus};
+use super::super::debug::*;
+
+// Macro to reduce repetitive debug prints
+macro_rules! file_debug {
+    ($msg:expr) => {
+        debug_print_str(concat!("File: ", $msg, "\n"));
+    };
+}
+
+const EFI_FILE_MODE_READ: u64 = 0x1;
+const KERNEL_PATH: &str = r"\EFI\BOOT\KERNEL.EFI";
+
+/// Fixed UTF-16 encode for KERNEL_PATH (no alloc).
+pub fn kernel_path_utf16() -> [u16; 32] {
+    // Enough for path + null
+    let path = KERNEL_PATH.encode_utf16().chain(core::iter::once(0u16));
+    let mut buf = [0u16; 32];
+    let mut i = 0;
+    for c in path {
+        if i < buf.len() - 1 {
+            buf[i] = c;
+            i += 1;
+        } else {
+            break;
+        }
+    }
+    buf[i] = 0; // Ensure null-term
+    buf
+}
+
+/// A RAII wrapper for EfiFile that automatically closes the file when it goes out of scope.
+pub struct EfiFileWrapper {
+    file: *mut EfiFile,
+}
+
+impl EfiFileWrapper {
+    pub fn new(file: *mut EfiFile) -> Self {
+        Self { file }
+    }
+}
+
+impl Drop for EfiFileWrapper {
+    fn drop(&mut self) {
+        if !self.file.is_null() {
+            // Safety:
+            // The `EfiFile` pointer is assumed to be valid from the `new` function.
+            // This is the last use of the pointer, so it is safe to dereference it for the `close` call.
+            let status = unsafe { ((*self.file).close)(self.file) };
+            if EfiStatus::from(status) != EfiStatus::Success {
+                // In a real application, you might want to log this error.
+                // For a bootloader, this might be a non-recoverable state.
+            }
+        }
+    }
+}
+
+/// Helper function to open a file from a directory handle.
+pub fn open_file(dir: &EfiFileWrapper, path: &[u16]) -> petroleum::common::Result<EfiFileWrapper> {
+    let mut file_handle: *mut EfiFile = ptr::null_mut();
+    let status = unsafe {
+        ((*dir.file).open)(
+            dir.file,
+            &mut file_handle,
+            path.as_ptr(),
+            EFI_FILE_MODE_READ,
+            0,
+        )
+    };
+    if EfiStatus::from(status) != EfiStatus::Success {
+        file_debug!("Failed to open file.");
+        return Err(BellowsError::FileIo("Failed to open file."));
+    }
+    file_debug!("Opened file.");
+    Ok(EfiFileWrapper::new(file_handle))
+}
+
+/// Read file content into memory
+pub fn read_file_to_memory(
+    bs: &EfiBootServices,
+    file: &EfiFileWrapper,
+) -> petroleum::common::Result<(usize, usize)> {
+    let mut file_info_buffer_size = 0;
+
+    let status = unsafe {
+        ((*file.file).get_info)(
+            file.file,
+            &EFI_FILE_INFO_GUID as *const _ as *const u8,
+            &mut file_info_buffer_size,
+            ptr::null_mut(),
+        )
+    };
+    if EfiStatus::from(status) != EfiStatus::BufferTooSmall || file_info_buffer_size == 0 {
+        file_debug!("Failed to get file info size.");
+        return Err(BellowsError::FileIo("Failed to get file info size."));
+    }
+
+    let mut file_info_buffer = [0u8; 4096];
+    if file_info_buffer_size > file_info_buffer.len() {
+        file_debug!("File info buffer too large.");
+        return Err(BellowsError::FileIo("File info buffer too large."));
+    }
+
+    let status = unsafe {
+        ((*file.file).get_info)(
+            file.file,
+            &EFI_FILE_INFO_GUID as *const _ as *const u8,
+            &mut file_info_buffer_size,
+            file_info_buffer.as_mut_ptr() as *mut c_void,
+        )
+    };
+    if EfiStatus::from(status) != EfiStatus::Success {
+        file_debug!("Failed to get file info.");
+        return Err(BellowsError::FileIo("Failed to get file info."));
+    }
+
+    let file_info: &EfiFileInfo = unsafe { &*(file_info_buffer.as_mut_ptr() as *const EfiFileInfo) };
+    let file_size = file_info.file_size as usize;
+
+    if file_size == 0 {
+        return Err(BellowsError::FileIo("Kernel file is empty."));
+    }
+
+    let pages = file_size.div_ceil(4096);
+    let mut phys_addr: usize = 0;
+
+    let status = (bs.allocate_pages)(
+        0usize,
+        petroleum::common::EfiMemoryType::EfiLoaderData,
+        pages,
+        &mut phys_addr,
+    );
+    if EfiStatus::from(status) != EfiStatus::Success {
+        file_debug!("Failed to allocate pages.");
+        return Err(BellowsError::AllocationFailed(
+            "Failed to allocate pages for kernel file.",
+        ));
+    }
+
+    let buf_ptr = phys_addr as *mut u8;
+    let mut read_size = file_size as u64;
+
+    let status = unsafe { ((*file.file).read)(file.file, &mut read_size, buf_ptr) };
+    if EfiStatus::from(status) != EfiStatus::Success || read_size as usize != file_size {
+        (bs.free_pages)(phys_addr, pages);
+        file_debug!("Failed to read file.");
+        return Err(BellowsError::FileIo(
+            "Failed to read kernel file or read size mismatch.",
+        ));
+    }
+
+    Ok((phys_addr, file_size))
+}
